+++ apps/dropper/src/utils/redis.js
@@ -1,5 +1,5 @@
 import { RedisClient, RedisUtil } from '@repo/common/classes';
-import { dropConfig } from '../src/config/config.js';
+import { dropConfig } from '../config/config.js';
 
 export const subRedisClient = new RedisClient(dropConfig.REDIS.REDIS_INFO).getClient();
